refactor(core): rename module emulator→load
For more consistency with so-called firehpc command.
